Move DrawLineSeparator to new GUIHelpers class
