+++ Packages/com.unity.inputsystem/InputSystem/Editor/Internal/GUIHelpers.cs
@@ -1,0 +1,49 @@
+#if UNITY_EDITOR
+using System;
+using UnityEditor;
+
+namespace UnityEngine.Experimental.Input.Editor
+{
+    internal static class GUIHelpers
+    {
+        public static class Styles
+        {
+            public static GUIStyle lineSeparator = new GUIStyle();
+
+            static Styles()
+            {
+                lineSeparator.fixedHeight = 1;
+                lineSeparator.margin.bottom = 2;
+                lineSeparator.margin.top = 2;
+            }
+        }
+
+        public static void DrawLineSeparator(string label = null)
+        {
+            var hasLabel = !string.IsNullOrEmpty(label);
+            EditorGUILayout.BeginVertical();
+            var rect = GUILayoutUtility.GetRect(GUIContent.none, Styles.lineSeparator, GUILayout.ExpandWidth(true));
+            var labelRect = new Rect();
+            GUIContent labelContent = null;
+            if (hasLabel)
+            {
+                labelContent = new GUIContent(label);
+                labelRect = GUILayoutUtility.GetRect(labelContent, EditorStyles.miniLabel, GUILayout.ExpandWidth(true));
+            }
+            EditorGUILayout.EndVertical();
+
+            if (Event.current.type != EventType.Repaint)
+                return;
+
+            var orgColor = GUI.color;
+            var tintColor = EditorGUIUtility.isProSkin ? new Color(0.12f, 0.12f, 0.12f, 1.333f) : new Color(0.6f, 0.6f, 0.6f, 1.333f);
+            GUI.color = GUI.color * tintColor;
+            GUI.DrawTexture(rect, EditorGUIUtility.whiteTexture);
+            GUI.color = orgColor;
+
+            if (hasLabel)
+                EditorGUI.LabelField(labelRect, labelContent, EditorStyles.miniLabel);
+        }
+    }
+}
+#endif // UNITY_EDITOR
